requirements.txt
fastapi
uvicorn[standard]
gunicorn
pytesseract
pdf2image
pillow
pymongo
python-multipart
openai
python-dotenv
streamlit
